Add high-level documentation:
* life of a computation
* tracing
* compilation
* execution

DO_NOT_SUBMIT: I would like feedback. Please be pedantic, it is important that the correct vocabulary is used and it is consistently applied.

Which sitemap organization is preferred? I am leaning to the second.
What is a good directory name? architecture?
Is there a place you think a code snippet would be useful?

DO_NOT_SUBMIT: Populate all the links.
DO_NOT_SUBMIT: Transform the links to be GitHub friendly.
PiperOrigin-RevId: 322607425